modifity assignOrdersController

+++ controllers/assignOrdersController.js
@@ -3,6 +3,19 @@ const pool = require('../db');
 const addAssignOrder = async (req, res) => {
     const { order_id, courier_id } = req.body;
     try {
+        const existOrder = await pool.query('SELECT * FROM orders WHERE id = $1', [order_id]);
+        if (existOrder.rows.length === 0) {
+            return res.status(404).json({ error: 'No existe la orden.' });
+        }
+        const existCourier = await pool.query('SELECT * FROM couriers WHERE id = $1', [courier_id]);
+        if (existCourier.rows.length === 0) {
+            return res.status(404).json({ error: 'Repartidor no encontrado' });
+        }
+        const existAssignment = await pool.query('SELECT * FROM assignment_order WHERE order_id = $1', [order_id]);
+        if (existAssignment.rows.length > 0) {
+            return res.status(400).json({ error: 'La orden ya ha sido asignada a un repartidor' });
+        }
+
         const result = await pool.query(
             'INSERT INTO assignment_order (order_id, courier_id) VALUES ($1, $2) RETURNING *',
             [order_id, courier_id]
@@ -20,7 +33,23 @@ const addAssignOrder = async (req, res) => {
 
 const getAssignOrders = async (req, res) => {
     try {
-        const result = await pool.query('SELECT * FROM assignment_order');
+        const result = await pool.query(`
+            SELECT 
+                ao.id AS assignment_id,
+                ao.assigned_at,
+                c.name AS courier_name,
+                c.phone AS courier_phone,
+                c.license_plate AS courier_license_plate,
+                o.id AS order_id,
+                o.total,
+                o.payment_method,
+                o.status
+            FROM assignment_order ao
+            INNER JOIN couriers c ON ao.courier_id = c.id
+            INNER JOIN orders o ON ao.order_id = o.id
+            ORDER BY ao.assigned_at DESC
+        `);
+
         res.status(200).json(result.rows);
     } catch (error) {
         console.error('Error obteniendo las asignaciones de orden:', error);
@@ -28,16 +57,35 @@ const getAssignOrders = async (req, res) => {
     }
 }
 
+
 const getAssignOrderByCourierId = async (req, res) => {
     const { courier_id } = req.params;
     try {
-        const result = await pool.query(
-            'SELECT * FROM assignment_order WHERE courier_id = $1',
-            [courier_id]
-        );
+        const existCourier = await pool.query('SELECT * FROM couriers WHERE id = $1', [courier_id]);
+        if (existCourier.rows.length === 0) {
+            return res.status(404).json({ error: 'Repartidor no encontrado' });
+        }
+
+        const result = await pool.query(`
+        SELECT 
+            ao.id AS assignment_id,
+            ao.assigned_at,
+            c.name AS courier_name,
+            c.phone AS courier_phone,
+            c.license_plate AS courier_license_plate,
+            o.id AS order_id,
+            o.total,
+            o.payment_method,
+            o.status
+        FROM assignment_order ao
+        INNER JOIN couriers c ON ao.courier_id = c.id
+        INNER JOIN orders o ON ao.order_id = o.id
+        WHERE ao.courier_id = $1
+        ORDER BY ao.assigned_at DESC
+    `,[courier_id]);
 
         if (result.rows.length === 0) {
-            return res.status(404).json({ error: 'No assignments found for this courier' });
+            return res.status(404).json({ error: 'No existe asignacion para este repartidor' });
         }
 
         res.status(200).json(result.rows);
@@ -50,13 +98,30 @@ const getAssignOrderByCourierId = async (req, res) => {
 const getAssignOrderByOrderId = async (req, res) => {
     const { order_id } = req.params;
     try {
-        const result = await pool.query(
-            'SELECT * FROM assignment_order WHERE order_id = $1',
-            [order_id]
-        );
+        const existOrder = await pool.query('SELECT * FROM orders WHERE id = $1', [order_id]);
+        if (existOrder.rows.length === 0) {
+            return res.status(404).json({ error: 'No existe la orden.' });
+        }
+        const result = await pool.query(`
+        SELECT 
+            ao.id AS assignment_id,
+            ao.assigned_at,
+            c.name AS courier_name,
+            c.phone AS courier_phone,
+            c.license_plate AS courier_license_plate,
+            o.id AS order_id,
+            o.total,
+            o.payment_method,
+            o.status
+        FROM assignment_order ao
+        INNER JOIN couriers c ON ao.courier_id = c.id
+        INNER JOIN orders o ON ao.order_id = o.id
+        WHERE ao.order_id = $1
+        ORDER BY ao.assigned_at DESC
+    `,[order_id]);
 
         if (result.rows.length === 0) {
-            return res.status(404).json({ error: 'No assignments found for this order' });
+            return res.status(404).json({ error: 'No existe asignación para esta orden' });
         }
 
         res.status(200).json(result.rows);
@@ -70,13 +135,21 @@ const updateAssignOrderCourier = async (req, res) => {
     const { order_id } = req.params;
     const { courier_id } = req.body;
     try {
+        const existOrder = await pool.query('SELECT * FROM orders WHERE id = $1', [order_id]);
+        if (existOrder.rows.length === 0) {
+            return res.status(404).json({ error: 'No existe la orden.' });
+        }
+        const existCourier = await pool.query('SELECT * FROM couriers WHERE id = $1', [courier_id]);
+        if (existCourier.rows.length === 0) {
+            return res.status(404).json({ error: 'Repartidor no encontrado' });
+        }
         const result = await pool.query(
             'UPDATE assignment_order SET courier_id = $1 WHERE order_id = $2 RETURNING *',
             [courier_id, order_id]
         );
 
         if (result.rows.length === 0) {
-            return res.status(404).json({ error: 'Assignment not found' });
+            return res.status(404).json({ error: 'No existe asignación para esta orden' });
         }
 
         res.status(200).json({
@@ -92,6 +165,11 @@ const updateAssignOrderCourier = async (req, res) => {
 const deleteAssignOrder = async (req, res) => {
     const { order_id } = req.params;
     try {
+        const existOrder = await pool.query('SELECT * FROM orders WHERE id = $1', [order_id]);
+        if (existOrder.rows.length === 0) {
+            return res.status(404).json({ error: 'No existe la orden.' });
+        }
+        
         const result = await pool.query(
             'DELETE FROM assignment_order WHERE order_id = $1 RETURNING *',
             [order_id]
